Reflect changes in callback prototype changed in oRTP.

+++ src/voip/audiostream.c
@@ -269,22 +269,22 @@ static OrtpRtcpXrPlcStatus audio_stream_get_rtcp_xr_plc_status(unsigned long use
 	return OrtpRtcpXrNoPlc;
 }
 
-static int8_t audio_stream_get_rtcp_xr_signal_level(unsigned long userdata) {
+static int audio_stream_get_rtcp_xr_signal_level(unsigned long userdata) {
 	AudioStream *stream = (AudioStream *)userdata;
 	if ((stream->features & AUDIO_STREAM_FEATURE_VOL_RCV) != 0) {
 		float volume;
 		ms_filter_call_method(stream->volrecv, MS_VOLUME_GET_MAX, &volume);
-		return (int8_t)volume;
+		return (int)volume;
 	}
 	return ORTP_RTCP_XR_UNAVAILABLE_PARAMETER;
 }
 
-static int8_t audio_stream_get_rtcp_xr_noise_level(unsigned long userdata) {
+static int audio_stream_get_rtcp_xr_noise_level(unsigned long userdata) {
 	AudioStream *stream = (AudioStream *)userdata;
 	if ((stream->features & AUDIO_STREAM_FEATURE_VOL_RCV) != 0) {
 		float volume;
 		ms_filter_call_method(stream->volrecv, MS_VOLUME_GET_MIN, &volume);
-		return (int8_t)volume;
+		return (int)volume;
 	}
 	return ORTP_RTCP_XR_UNAVAILABLE_PARAMETER;
 }
